backend/pyproject.toml
[project]
name = "pixelscale-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.42.15",
    "fastapi[standard]>=0.127.0",
    "pillow>=12.0.0",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.12.0",
    "python-multipart>=0.0.21",
    "sqlalchemy>=2.0.45",
]
